+++ src/main/java/rbasamoyai/createbigcannons/datagen/recipes/CBCCraftingRecipeProvider.java
@@ -488,6 +488,15 @@ public abstract class CBCCraftingRecipeProvider extends RecipeProvider {
 			.unlockedBy("has_wool", has(ItemTags.WOOL))
 			.save(cons);
 
+        ShapedRecipeBuilder.shaped(RecipeCategory.TOOLS, CBCItems.BLOCK_ARMOR_INSPECTION_TOOL)
+            .define('M', AllItems.PRECISION_MECHANISM).define('S', CBCItems.RECOIL_SPRING).define('P', Items.IRON_BARS)
+            .pattern("M")
+            .pattern("S")
+            .pattern("P")
+            .unlockedBy(getHasName(CBCItems.RECOIL_SPRING), has(CBCItems.RECOIL_SPRING))
+            .unlockedBy(getHasName(AllItems.PRECISION_MECHANISM), has(AllItems.PRECISION_MECHANISM))
+            .save(cons);
+
 		specialRecipe(CBCRecipeTypes.MUNITION_FUZING.getSerializer(), cons, "munition_fuzing");
 		specialRecipe(CBCRecipeTypes.CARTRIDGE_ASSEMBLY.getSerializer(), cons, "cartridge_assembly");
 		specialRecipe(CBCRecipeTypes.BIG_CARTRIDGE_FILLING.getSerializer(), cons, "big_cartridge_filling");